TLS, better interface, netsplit-resilience, but IDENTIFY doesn't work yet!
Looks like timing is critical during IDENTIFY with NickServ. This code fails
to identify because IDENTIFY is called to early - if you put a breakpoint in
privatemsg() and call IDENTIFY from there, it works.
So, I've tried to defer the IDENTIFY phase by scheduling it with the reactor,
but it doesn't seem to work either (it would be strange to have to wait more
than 5 seconds).